Fix for initial packet CRC errors not supplying battery info. Only works once (2 crc errors still fails).
git-svn-id: https://subversion.assembla.com/svn/etl_webapp/trunk@231 3fa7648a-587c-4a36-b6a6-96aee5749d1d

+++ iOS/ETL-interface-test/ETL-interface-test/ETLProgramViewController.h
@@ -21,7 +21,7 @@
     
     NSTimer * progressBarTimer, *startTimer;
     
-    uint8_t batteryLevel;
+    uint16_t batteryLevel;
 }
 
 - (void)didFinishProgramming:(id)sender;
